% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert}
\alias{assert}
\alias{assert_non_na}
\alias{assert_non_na_and_unique}
\title{Assert vector characteristics}
\usage{
assert_non_na(x, class_vector, proportion_minimum)
assert_non_na_and_unique(x, class_vector)
}
\arguments{
\item{x}{Vector to inspect.  Required.}

\item{class_vector}{The required \code{\link[base:class]{base::class()}} of the vector.  If the
parameter is missing, the object's class is not checked.}

\item{proportion_minimum}{The (inclusive) minimum proportion of the
vector's elements that should meet the requirement. If missing, all
elements must pass.}
}
\description{
Assert a vector meets important data-quality requirements.
}
\examples{
requireNamespace("OuhscMunge")
OuhscMunge::assert_non_na(1:30, "integer")
\dontrun{
OuhscMunge::assert_non_na(c(1:30, NA_integer_), "integer")
}
}
