% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert-version.R
\name{assert_version}
\alias{assert_version}
\alias{assert_version_r}
\alias{assert_version_r,}
\alias{assert_version_package}
\title{I that the local machine is using an acceptable version.}
\usage{
assert_version_r(minimum = base::package_version("4.2.1"))
}
\arguments{
\item{minimum}{A \link{package_version} or \link{character} that specifies
the version of the software being examined
(ie, R, a package, or an ODBC driver).}
}
\value{
An error if the minimum version is not met.
If the local machine is using an acceptable version, an \code{invisible()} \code{TRUE}
is returned.
}
\description{
Assert that the local machine is using a version that satisfies
the minimum version specified.
}
\note{
These functions help us assert the the local machine has an acceptable
version of the software running.

For \code{\link[=assert_version_r]{assert_version_r()}}, the current default value is "4.2.0" because
it introduced the
\href{https://davidbudzynski.github.io/general/2022/04/23/r-native-placeholder.html}{placeholder}
for the native pipe.  Future versions of OuhscMunge will likely increase
the default value to keep pace with important developments to R.
}
\examples{
# Check R
assert_version_r("3.1.0")
assert_version_r()
# Fails: assert_version_r("99.1.0")

# Check packages
assert_version_package("base", "3.1.0")
assert_version_package("OuhscMunge", "0.1.0")
# Fails: assert_version_package("base", "99.1.0")
# Fails: assert_version_package("OuhscMunge", "9.1.0")
# Fails:
# Fails: assert_version_package("OuhscMungeee", "9.1.0")


}
\author{
Will Beasley
}
