% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{clump_date}
\alias{clump_date}
\alias{clump_month_date}
\alias{clump_week_date}
\title{Assign date for a given year & month}
\usage{
clump_month_date( date_detailed, day_of_month=15L )
clump_week_date( date_detailed, day_of_week=2L )
}
\arguments{
\item{date_detailed}{The \code{Date} value containing the desired year and month.  The day will be overwritten. Required}

\item{day_of_month}{The factor label assigned to the missing value.  Defaults to \code{15} (\emph{i.e.}, the middle of the month).}

\item{day_of_week}{The factor label assigned to the missing value.  Defaults to \code{2} (\emph{i.e.}, Monday).}
}
\value{
An array of \code{Date} values.
}
\description{
This accepts a date, but changes the day.  Set/degrade/clump all the days within a month/week to the same day.
}
\details{
We use this frequently to set/degrade/clump all the days to the middle of their respective month (ie, the 15th day).
The midpoint of a month is usually the most appropriate summary location.  It makes graphs more intuitive.
Using the midpoint of month can also avoid problems with timezones.  A date won't get nudged to a neighboring month accidentally.
}
\note{
A \code{stop} error will be thrown if \code{date_detailed} is not a \code{Date}.  \code{day_of_month} must be bounded by \verb{[1, 31]}, and
\code{day_of_week} must be bounded by \verb{[1, 7]}.
Be careful that if you set a November date the 31st day, the result will be December 1st.
Consequently, we recommend not setting the day to a value after the 28.

The \href{https://docs.microsoft.com/en-us/sql/t-sql/functions/datepart-transact-sql}{sql equivalent} to \code{clump_month_date()}
is \verb{CAST(convert(char(7), GETDATE(), 126) + '-15' AS date)}.
The \href{https://blog.sqlauthority.com/2007/08/20/sql-server-find-monday-of-the-current-week/}{sql equivalent} to \code{clump_week_date()}
is \verb{SELECT DATEADD(wk, DATEDIFF(wk,0,GETDATE()), 0)}
}
\examples{
library(OuhscMunge)
detailed <- seq.Date(from=as.Date("2011-04-21"), to=as.Date("2011-07-14"), by="day")

clumped_month <- clump_month_date(detailed)
table(clumped_month)
# 2011-04-15 2011-05-15 2011-06-15 2011-07-15
#         10         31         30         14

clumped_week <- clump_week_date(detailed)
table(clumped_week)
# 2011-04-18 2011-04-25 2011-05-02 2011-05-09 2011-05-16 2011-05-23 2011-05-30
#          3          7          7          7          7          7          7
# 2011-06-06 2011-06-13 2011-06-20 2011-06-27 2011-07-04 2011-07-11
#          7          7          7          7          7          5
}
\seealso{
These functions are gloves around \code{\link[lubridate:day]{lubridate::day()}} and
\code{\link[lubridate:day]{lubridate::wday()}}

Essentially the add just error-checking and default values.
}
\author{
Will Beasley
}
