% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{clump_month_date}
\alias{clump_month_date}
\title{Assign date for a given year & month}
\usage{
clump_month_date(date_detailed, day_of_month = 15L)
}
\arguments{
\item{date_detailed}{The \code{Date} value containing the desired year and month.  The day will be overwritten. Required}

\item{day_of_month}{The factor label assigned to the missing value.  Defaults to \code{15}.}
}
\value{
An array of \code{Date} values.
}
\description{
This accepts a date, but changes the day.  Set/degrade/clump all the days within a month to the same day.
}
\details{
We use this frequently to set/degrade/clump all the days to the middle of their respective month (ie, the 15th day).  
The midpoint of a month is usually the most appropriate summary location.  It makes graphs more intuitive.  
Using the midpoint of month can also avoid problems with timezones.  A date won't get nudged to a neighboring month accidentally.
}
\note{
A \code{stop} error will be thrown if \code{date_detailed} is not a \code{Date}, or if \code{day_of_month} is not bounded by \code{[1, 31]}.
Be careful that if you set a November date the 31st day, the result will be December 1st.  
Consequently, we recommend not setting the day to a value after the 28.
}
\examples{
library(OuhscMunge)
detailed <- seq.Date(from=as.Date("2011-04-21"), to=as.Date("2011-07-14"), by="day")
clumped <- clump_month_date(detailed)
table(clumped)
# 2011-04-15 2011-05-15 2011-06-15 2011-07-15  
#         10         31         30         14  
}
\author{
Will Beasley
}
