% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut-wth-nas.R
\name{cut_with_nas}
\alias{cut_with_nas}
\title{Convert numeric to factor, with an explicit level for \code{NA}s}
\usage{
cut_with_nas(x, .missing = "Unknown", ...)
}
\arguments{
\item{x}{A \code{numeric} or \code{integer} vector to cut into factor levels.  Required.}

\item{.missing}{The name of the level representing the \code{NA} values within \code{x}.}

\item{...}{further arguments passed to \code{\link[base:cut]{base::cut()}}.}
}
\value{
A \code{factor}.
}
\description{
Like \code{\link[base:cut]{base::cut()}}, but creates a level representing the missing values.
}
\note{
Discussed in the Stack Overflow question,
\href{https://stackoverflow.com/questions/50882620/cut-a-variable-with-missing-values}{"cut() a variable with missing values"}
}
\examples{
w      <- c(0L, NA_integer_, 22:25, NA_integer_, 40)
breaks <- c(0, 25, 50)
labels <- c("lower", "upper")

cut_with_nas(w, breaks=2)
cut_with_nas(w, breaks=breaks, labels=labels)
cut_with_nas(w, breaks=breaks               )
cut_with_nas(w, breaks=breaks               , include.lowest=T)
cut_with_nas(w, breaks=breaks               , include.lowest=T, right=F)
cut_with_nas(w, breaks=breaks                                 , right=F)
}
\author{
Will Beasley
}
