% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-frame-compare.R
\name{data_frame_compare}
\alias{data_frame_compare}
\title{Identify new rows and add.}
\usage{
data_frame_compare(d_original, d_current, columns)
}
\arguments{
\item{d_original}{A \code{data.frame} that serves as the existing metadata file
that potentialy needs to be updated.  Required.}

\item{d_current}{A \code{data.frame} that contains records potentialy missing from
\code{d_original}. Required.}

\item{columns}{A vector to describe.}
}
\value{
`data.frame``
}
\description{
These functions help to compare two metadata frames as assess if new rows
should be added.
}
\examples{
library("magrittr")
ds_original <- tibble::tibble(
  x1  = c(1, 3, 4),
  x2  = letters[c(1, 3, 4)],
  x3  = c(11, 13, 14)
)

ds_current <- tibble::tibble(
  x1  = c(1:5, 1, 5, 5),
  x2  = c(letters[1:5], "a", "e", "e"),
  x3  = c(11, 12, 13, 14, 15, 11, 15, 17)
)

ds_new_1 <-
  ds_current \%>\%
  dplyr::anti_join(ds_original, by = c("x1", "x2"))
ds_new_1
ds_new_2 <- data_frame_compare(ds_original, ds_current, c("x1", "x2"))
ds_new_2

}
\author{
Will Beasley
}
