% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-frame-stack.R
\name{data_frame_stack}
\alias{data_frame_stack}
\title{Verify equivalent structure of two dataset}
\usage{
data_frame_stack(d_original, d_current, keys)
}
\arguments{
\item{d_original}{A \code{data.frame} that serves as the existing metadata file
that potentialy needs to be updated.  Required.}

\item{d_current}{A \code{data.frame} that contains records potentialy missing from
\code{d_original}. Required.}

\item{keys}{Column names that represent unique combination. \code{character} vector. Optional.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble}} that combines \code{d_original} with the new records from \code{d_current}.
}
\description{
These functions help to compare two metadata frames as assess if new rows
should be added.
}
\note{
each dataset is verified to not have more then one
row with the same values in the combination of \code{keys}
}
\examples{
library("magrittr")
ds_original <- tibble::tibble(
  x1  = c(1, 3, 4),
  x2  = letters[c(1, 3, 4)],
  x3  = c(11, 13, 14)
)

ds_current <- tibble::tibble(
  x1  = c(1:5, 1, 5),
  x2  = c(letters[1:5], "x", "y"),
  x3  = c(11, 12, 13, 14, 15, 11, 15)
)

data_frame_stack(ds_original, ds_current, c("x1", "x2"))

ds_current \%>\%
  dplyr::anti_join(ds_original, by = c("x1", "x2"))

}
\seealso{
\code{\link[=data_frame_compare_structure]{data_frame_compare_structure()}}
}
\author{
Will Beasley
}
