% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.R
\name{date_range}
\alias{date_range}
\alias{date_min_with_nas}
\alias{date_max_with_nas}
\alias{date_range_with_nas}
\alias{min_with_nas_numeric}
\alias{max_with_nas_numeric}
\alias{range_with_nas_numeric}
\alias{min_with_nas_integer}
\alias{max_with_nas_integer}
\alias{range_with_nas_integer}
\title{Find date ranges in the  prescence of missing subsets}
\usage{
date_min_with_nas(x)
date_max_with_nas(x)
date_range_with_nas(x)
min_with_nas_numeric(x)
max_with_nas_numeric(x)
range_with_nas_numeric(x)
min_with_nas_integer(x)
max_with_nas_integer(x)
range_with_nas_integer(x)
}
\arguments{
\item{x}{The input date vector.  Required}
}
\value{
A date value, that's possibly \code{NA}.
}
\description{
Return \code{NA} for the min and max of a date vector if no nonmissing values are presence
}
\note{
This function is a workaround for a weakness in \code{base::min.date()} and \code{base::max.date()}.
If no nonmissing values are present, both functions \emph{return} +/-\code{Inf}, but \emph{print} \code{NA}.
These two function return and print \code{NA}, which behaves like SQL
(and probably matches the expectations of most users).

See Stack Overflow Questions \href{https://stackoverflow.com/questions/48470746/using-dplyrgroup-by-to-find-min-dates-with-nas}{Using dplyr::group_by() to find min dates with NAs}
and \href{https://stackoverflow.com/questions/27554410/r-inf-when-it-has-class-date-is-printing-na}{R Inf when it has class Date is printing NA}.

The foundation of these functions was proposed in a \href{https://stackoverflow.com/a/48471923/1082435}{response}
by Edward Visel (SO username \href{https://stackoverflow.com/users/4497050/alistaire}{alistaire}).
}
\examples{
library(OuhscMunge)
date_min_with_nas(c(NA, NA, NA))
date_min_with_nas(as.Date(NA_character_))
date_min_with_nas(as.Date(character(0)))
date_min_with_nas(as.Date(c("2009-04-21", "2017-12-27", NA_character_)))
}
\author{
Edward Visel, Will Beasley
}
