% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deterge.R
\name{deterge}
\alias{deterge}
\alias{deterge_to_double}
\alias{deterge_to_integer}
\title{Convert (and possibly clean) a vector}
\usage{
deterge_to_double(x, bound_lower, bound_upper)
deterge_to_integer(x, bound_lower, bound_upper)
}
\arguments{
\item{x}{The input vector that needs to be cast.  Required}

\item{bound_lower}{Elements below this inclusive threshold will be set to \code{NA}.}

\item{bound_upper}{Elements above this inclusive threshold will be set to \code{NA}.}
}
\value{
An array of values.
}
\description{
Cast values to desired data type.
}
\details{
--write something here--
}
\examples{
library(OuhscMunge)
deterge_to_double(c(NA, 1:10), 4, 8)
deterge_to_integer(c(NA, 1:10), 4, 8)
}
\author{
Will Beasley
}

