% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute-sql-file.R
\name{execute_sql_file}
\alias{execute_sql_file}
\title{Execute a SQL file}
\usage{
execute_sql_file(
  path_sql,
  dsn,
  execute = TRUE,
  minimum_row_count = 0L,
  timezone = "UTC",
  timezone_out = "UTC"
)
}
\arguments{
\item{path_sql}{A vector to of names to convert.  Required character.}

\item{dsn}{The name of a \href{https://en.wikipedia.org/wiki/Data_source_name}{DSN} defined on your local machine  Required character.}

\item{execute}{Indicates if \code{DBI::dbExecute()} should be used
(which typically returns a scalar).  Otherwise, \code{DBI::dbGetQuery()} is used,
(which will return a \code{tibble::tibble}).  Required \code{logical}.}

\item{minimum_row_count}{If \code{execute} is false, the returned dataset should have at least this many rows, or an error will be thrown.  Default of 0.  Required integer.}

\item{timezone}{The server time zone.  Passed to \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}

\item{timezone_out}{The time zone returned to R. Passed to \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.  See https://www.tidyverse.org/blog/2019/12/odbc-1-2-0/.}
}
\value{
A vector of converted names.
}
\description{
Read a SQL file, and execute its text using the \code{odbc} and \code{DBI} packages.
}
\examples{
\dontrun{
execute_sql_file("inst/hdid-select.sql", "cdw_cache_staging")
execute_sql_file("inst/condense-date.sql", "cdw_cache_staging")
}
}
\author{
Will Beasley
}
