% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first-nonmissing.R
\name{first_nonmissing}
\alias{first_nonmissing}
\title{First nonmissing element in a vector}
\usage{
first_nonmissing(x)
}
\arguments{
\item{x}{A vector to of names to convert.}
}
\value{
A vector of converted names.
}
\description{
Take the first value that isn't missing.  Adapted from http://stackoverflow.com/a/40515261/1082435.
}
\examples{
first_nonmissing(c(NA, "b", "c"))
first_nonmissing(c(NA_character_, NA_character_))
first_nonmissing(character(0))
}
\author{
Will Beasley
}
