% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash-and-salt.R
\name{hash_and_salt_sha_256}
\alias{hash_and_salt_sha_256}
\title{Salt and hash a value}
\usage{
hash_and_salt_sha_256(
  x,
  salt = "",
  min_characters = 1L,
  max_characters = 2048L,
  na_if = c("")
)
}
\arguments{
\item{x}{A vector of values to convert.  it should be a \code{character} vector, or something that can be cast to a \code{character} vector.}

\item{salt}{A single-element character vector.}

\item{min_characters}{The minimum count of characters that \code{x} is allowed to be.  Must be an \code{integer} or \code{numeric} data type.}

\item{max_characters}{The maximum count of characters that \code{x} is allowed to be.  Must be an \code{integer} or \code{numeric} data type.}

\item{na_if}{A vector of characters that should produce a has of \code{NA_character_}.  Default of \code{c("")}.}
}
\value{
A character vector.
}
\description{
Uses \code{\link[digest:digest]{digest::digest()}} to hash as (salted) value, using 'SHA-256'.
If the \code{x} isn't already a \code{character} vector, it's converted to one (even if \code{x} inherits from character).

This approach protects the actual value of \code{x}, while still allowing a downstream user
to determine which cells were derived from the same \code{x}.

For example, suppose a patient's
\href{https://www.ahrq.gov:443/data/ushik.html}{mrn}
of '111' is hashed, and the output is 'abc'.
(To view it's real value, execute \code{OuhscMunge::hash_and_salt_sha_256(111)}.)
When given the value 'abc', is it computational infeasible to determine the input
had been '111' (especially when \emph{salted}).  However, when you see that two visits
have an mrn of 'abc', you can determine the same patient generated both visits.
}
\examples{
x    <- letters[1:5]

salt <- "abc123"
hash_and_salt_sha_256(x, salt)

# If an unsalted hash is desired, leave the `salt` parameter blank
hash_and_salt_sha_256(x)

# By default, a zero-length character produces hash of NA.
hash_and_salt_sha_256(c("a", "", "c"))
}
\references{
\href{https://en.wikipedia.org/wiki/Cryptographic_hash_function}{cryptographic hash function}
and \href{https://en.wikipedia.org/wiki/Salt_(cryptography)}{salts}.
}
\author{
Will Beasley
}
