% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-packages-addin.R
\name{install_packages_addin}
\alias{install_packages_addin}
\title{Install important packages}
\usage{
install_packages_addin()
}
\description{
Installs the \href{https://github.com/OuhscBbmc/RedcapExamplesAndPatterns/blob/master/utility/package-dependency-list.csv}{important packages}
typically used by \href{https://github.com/OuhscBbmc}{BBMC} data analysts.
}
