% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join-stats.R
\name{match_statistics}
\alias{match_statistics}
\title{Create explicit factor level for missing values.}
\usage{
match_statistics(d_parent, d_child, join_columns)
}
\arguments{
\item{d_parent}{A \code{data.frame} of the parent table.}

\item{d_child}{A \code{data.frame} of the child table.}

\item{join_columns}{The \code{character} vector of the column names used to join to parent and child tables.}
}
\value{
A \code{numeric} array of the following elements: \code{in_child}, \code{not_in_child}, \code{deadbeat_proportion}, \code{in_parent}, \code{not_in_parent}, and \code{orphan_proportion}.
}
\description{
Missing values are converted to a factor level.  This explicit assignment can reduce the chances that missing values are inadvertantly ignored.  
It also allows the presence of a missing to become a predictor in models.
}
\author{
Will Beasley
}

