% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open-dsn-channel.R
\name{open_dsn_channel_sqls}
\alias{open_dsn_channel_sqls}
\title{Open an ODBC channel to a SQL Server database}
\usage{
open_dsn_channel_sqls(dsn_name,
  driver_version_minimum = numeric_version("13.0"))
}
\arguments{
\item{dsn_name}{Name of the locally-defined DSN passed to \url{RODBC::odbcConnect()}.}

\item{driver_version_minimum}{Represnted as a \code{\link[base:numeric_version]{base::numeric_version()}}}
}
\description{
Creates & opens a channel and checks its important characteristics.
}
\details{
A DSN channel requires more code than usual to diagnose problems, because the DSN
is defined on the local computer, and is not under the control of the repository.
This function wraps the basic \url{RODBC::odbcConnect()} function with some
checks.  If unsuccessful, it returns some hints how to correct the problem, such as downloading
the newest version from the \href{https://docs.microsoft.com/en-us/sql/connect/odbc/download-odbc-driver-for-sql-server}{Microsoft website}.
}
\examples{
\dontrun{
requireNamespace("OuhscMunge") 

OuhscMunge::open_dsn_channel_sqls(
  dsn_name        = "miechv_eval"
)
}
}
