% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OuhscMunge-package.r
\docType{package}
\name{OuhscMunge-package}
\alias{OuhscMunge-package}
\title{OuhscMunge: Data Manipulation Operations}
\description{
Data manipulation operations frequently used in OUHSC BBMC projects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OuhscBbmc/OuhscMunge}
  \item \url{http://ouhsc.edu/bbmc/}
  \item Report bugs at \url{https://github.com/OuhscBbmc/OuhscMunge/issues}
}

}
\author{
\strong{Maintainer}: Will Beasley \email{wibeasley@hotmail.com} (\href{https://orcid.org/0000-0002-5613-5006}{ORCID})

}
\keyword{internal}
