% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readr-spec-aligned.R
\name{readr_spec_aligned}
\alias{readr_spec_aligned}
\title{Align & qualify the \code{col_types} specification passed to \code{\link[readr:read_delim]{readr::spec_csv()}}.}
\usage{
readr_spec_aligned(...)
}
\arguments{
\item{...}{values passed to \code{\link[readr:spec_delim]{readr::spec_csv()}}.}
}
\value{
Prints formatted code to the console.
}
\description{
These functions are used during the execution of a program.  Rather they produce snippets
that can be pasted into code, and help the developer avoid some typing.
}
\examples{
# The aligned values produced by this funciton.
readr_spec_aligned(system.file("test-data/subject-1.csv", package = "OuhscMunge"))
readr_spec_aligned(system.file("package-dependency-list.csv", package = "OuhscMunge"))

# For reference, the unaligned output of `readr::spec_csv()`.
readr::spec_csv(system.file("test-data/subject-1.csv", package = "OuhscMunge"))
readr::spec_csv(system.file("package-dependency-list.csv", package = "OuhscMunge"))
}
\author{
Will Beasley
}
