% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace-with-nas.R
\name{replace_with_nas}
\alias{replace_with_nas}
\title{Convert blank, zero-length values to \code{NA}s for a variety of data types.}
\usage{
replace_with_nas(x, return_type = NULL)
}
\arguments{
\item{x}{An array of values. Required}

\item{return_type}{Data type of returned vector.  Optional}
}
\value{
An array of values with \code{NA}s.
}
\description{
Elements of zero-length are converted to \code{NA}s.  Can force
cohersion to an optionally-specified data type.
}
\details{
If \code{return_type} is missing, returned data type will match input.
Supports cohersion to \code{integer}, \code{numeric}, \code{character}, and \code{Date} vectors.
}
\note{
Contact the package author if you'd like the function generalized so that additional values
(other that \code{""}) are converted to \code{NA}s.
}
\examples{
library(OuhscMunge) #Load the package into the current R session.
replace_with_nas(c("a", "b", "", "d", ""))
replace_with_nas(c("a", "b", "", "d", ""), return_type="character")

replace_with_nas(c(1, 2, "", "", 5), return_type="character")
replace_with_nas(c(1, 2, "", "", 5)) #Equivalent to previous line.
replace_with_nas(c(1, 2, "", "", 5), return_type="integer")
replace_with_nas(c(1, 2, "", "", 5), return_type="numeric")

replace_with_nas(c("2011-02-03", "", "", "2011-02-24"), return_type="Date")
}
\author{
Will Beasley
}

