% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve-key-value.R
\name{retrieve_key_value}
\alias{retrieve_key_value}
\title{Read a value stored in a database.}
\usage{
retrieve_key_value(key, project_name, dsn_name, channel = NULL)
}
\arguments{
\item{key}{The key associated with the desired value.  Required character vector with one element}

\item{project_name}{The project name associated with the desired value.  Required character vector with one element}

\item{dsn_name}{Name of the locally-defined DSN passed to \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}

\item{channel}{An \emph{optional} connection handle as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.  See Details below. Optional.}
}
\value{
A \code{character} vector with one element.
}
\description{
Facilitates retrieval of key-value pairs that are stored in a SQL Server database.
}
\details{
The database table and stored procedure must defined as:

\if{html}{\out{<div class="sourceCode sql">}}\preformatted{CREATE TABLE security_private.tbl_key_value_static(
  id                     smallint     identity(1,1) primary key,
  project                varchar(50)  not null,
  attribute              varchar(90)  not null,
  value                  varchar(200) not null,
  file_last_updated_date date         not null,
  retired                bit          not null
)

CREATE PROCEDURE security.prc_key_value_static
  @project varchar(50),
  @attribute varchar(90)
AS
BEGIN
  SET NOCOUNT ON;
  SELECT value from security_private.tbl_key_value_static
  WHERE project = @project and attribute = @attribute
END
}\if{html}{\out{</div>}}
}
\note{
Currently only the 'static' key-value pairs are retrieved through this function.
Talk to Will if you need to retrieve the 'rolling' or the 'personal' key-value pairs.
}
\examples{
\dontrun{
value <- retrieve_key_value("file-server", "bbmc", "BbmcSecurity")
}
}
\author{
Will Beasley
}
