% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row.R
\name{row_sum}
\alias{row_sum}
\title{Find the sum of selected columns within a row}
\usage{
row_sum(
  d,
  columns_to_average = character(0),
  pattern,
  new_column_name = "row_sum",
  threshold_proportion = 0.75,
  verbose = FALSE
)
}
\arguments{
\item{d}{The data.frame containing the values to sum. Required.}

\item{columns_to_average}{A character vector containing the columns
names to sum.
If empty, \code{pattern} is used to select columns. Optional.}

\item{pattern}{A regular expression pattern passed to \code{\link[base:grep]{base::grep()}}
(with \code{perl = TRUE}).  Optional}

\item{new_column_name}{The name of the new column that represents the sum
of the specified columns.  Required.}

\item{threshold_proportion}{Designates the minimum proportion of columns
that have a nonmissing values (within each row) in order to return a sum.
Required; defaults to to 0.75.}

\item{vebose}{a logical value to designate if extra information is
displayed in the console,
such as which columns are matched by \code{pattern}.}
}
\value{
The data.frame \code{d}, with the additional column containing the row sum.
}
\description{
Sums across columns within a row,
while accounting for nonmissingness.
Specify the desired columns by passing their explicit column names or
by passing a regular expression to matches the column names.
}
\details{
If the specified columns are all logicals or integers,
the new column will be an \link{integer}.
Otherwise the new column will be a \link{double}.
}
\examples{
library(OuhscMunge) #Load the package into the current R session.

}
\author{
Will Beasley
}
