% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snake-case.R
\name{snake_case}
\alias{snake_case}
\title{Convert variable names to snake_case}
\usage{
snake_case(x)
}
\arguments{
\item{x}{A vector of names to convert.}
}
\value{
A vector of converted names.
}
\description{
This function attempts to convert variables to snake_case, even if it's already in snake_case.
The important regex lines were posted by Stack Overflow user \href{https://stackoverflow.com/users/129879/epost}{epost}
in \href{https://stackoverflow.com/a/1176023/1082435}{"Elegant Python function to convert CamelCase to snake_case?"}.
}
\note{
This series of regexes has an advantages over the current
implementations of \href{https://github.com/decisionpatterns/lettercase}{\code{lettercase::str_snake_case()}}
and \href{https://tazinho.github.io/snakecase/reference/caseconverter.html}{\code{snakecase::to_snake_case()}}.
The former converts "PatientDOB" to "patientdob" and the latter converts
"patient.dob" to "patient_._dob".  I'll keep an eye on these packages
(\emph{i.e.}, \href{https://github.com/decisionpatterns/lettercase/issues/1}{lettercase #1} for 'camelCase'
and \href{https://github.com/Tazinho/snakecase/issues/101}{snakecase #101}).  I'd prefer
to use one of them, instead of maintaining the functions.
}
\examples{
snake_case(colnames(datasets::OrchardSprays))
snake_case(colnames(datasets::iris))
snake_case(c("PatientID", "PatientDOB", "DOB", "name.last", "name.first"))
}
\author{
Will Beasley
}
