% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify-data-frame.R
\name{verify_data_frame}
\alias{verify_data_frame}
\title{Object inherits from \href{base::data.frame()}{data.frame}.}
\usage{
verify_data_frame(d, minimum_row_count = 10L)
}
\arguments{
\item{d}{The object to verify.  Required.}

\item{minimum_row_count}{The \code{data.frame} should have at least this many rows.  Defaults to 10.
The datatype does not have to be an \code{integer}, but should be safely convertible to an integer.}
}
\description{
Check that the object inherits from \href{base::data.frame()}{data.frame}.
If not, throw an error.

This function will be deprecated in the future.  If you're developing new code,
consider the superior \url{checkmate} functions, \code{\link[checkmate:assert_data_frame]{checkmate::assert_data_frame()}} and \code{\link[checkmate:assert_tibble]{checkmate::assert_tibble()}}

This helps check database-reading functions (\emph{e.g.}, \code{\link[RODBC:sqlQuery]{RODBC::sqlQuery()}}) that return a \code{data.frame}
if successful, and a \code{character} vector is unsucessful.

A minimum row count can be used to check that a trivially small number of rows was not returned.
If \code{minimum_row_count} is set to 0, the function is similar to \code{checkmate::assert_class(d, "data.frame")},
but with with a more specific error message.
}
\examples{
verify_data_frame(datasets::OrchardSprays, 20)
verify_data_frame(datasets::iris, 4)
}
\seealso{
\code{\link[checkmate:assert_class]{checkmate::assert_class()}}
}
\author{
Will Beasley
}
