% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-frame-uniqueness.R
\name{data_frame_test_uniqueness}
\alias{data_frame_test_uniqueness}
\alias{data_frame_assert_uniqueness}
\title{Check uniqueness of column combinations}
\usage{
data_frame_test_uniqueness(d, keys, display_count = 10L)

data_frame_assert_uniqueness(d, keys, display_count = 10L)
}
\arguments{
\item{d}{A \code{data.frame} to examine.  Required.}

\item{keys}{A \code{character} vector specifying the (combination of) columns that should be unique.}

\item{display_count}{Maximum number of uniquness violations to display.}
}
\value{
A \code{logical} value indicating if uniqueness is satisfied.
If \code{FALSE}, the top rows are printed to the console.
}
\description{
Useful checking that duplicates of primary keys don't exist
}
\examples{
data_frame_test_uniqueness(mtcars, c("cyl"))
data_frame_test_uniqueness(mtcars, c("cyl", "vs"))
data_frame_test_uniqueness(mtcars, c("cyl", "hp"))
data_frame_test_uniqueness(mtcars, c("cyl", "hp"), display_count=0)
data_frame_test_uniqueness(mtcars, c("mpg", "wt"))

\dontrun{
data_frame_assert_uniqueness(mtcars, c("cyl"))
data_frame_assert_uniqueness(mtcars, c("cyl", "vs"))
data_frame_assert_uniqueness(mtcars, c("mpg", "wt"))
}

}
\author{
Will Beasley
}
