% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute-sql-file.R
\name{execute_sql_file}
\alias{execute_sql_file}
\title{Execute a SQL file}
\usage{
execute_sql_file(path_sql, dsn)
}
\arguments{
\item{path_sql}{A vector to of names to convert.  Required character.}

\item{dsn}{The name of a \href{https://en.wikipedia.org/wiki/Data_source_name}{DSN} defined on your local machine  Required character.}
}
\value{
A vector of converted names.
}
\description{
Read a SQL file, and execute its text using the \code{odbc} and \code{DBI} packages.
}
\examples{
\dontrun{
execute_sql_file("inst/hdid-select.sql", "cdw_cache_staging")
execute_sql_file("inst/condense-date.sql", "cdw_cache_staging")
}
}
\author{
Will Beasley
}
