% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first-nonmissing.R
\name{first_nonmissing}
\alias{first_nonmissing}
\title{First nonmissing element in a vector}
\usage{
first_nonmissing(x, value_if_all_na = NULL, na_codes = NULL)
}
\arguments{
\item{x}{A vector of values to collapse.}

\item{value_if_all_na}{A scalar value to return if all elements are missing
(\emph{i.e.}, they are all either \code{NA}, or one of the \code{na_codes}).}

\item{na_codes}{A vector of codes that represent missing values.}
}
\value{
A scalar of converted names.
}
\description{
Take the first value that isn't missing.  Adapted from http://stackoverflow.com/a/40515261/1082435.
}
\details{
\code{value_if_all_na} and \code{na_codes}must have the same data type as \code{x}.

If \code{value_if_all_na} is null, then an \code{NA} will be returned.
If \code{na_codes} is null, then all non-NA values are considered.
}
\examples{
first_nonmissing(c(NA, "b", "c"))
first_nonmissing(c(NA_character_, NA_character_))
first_nonmissing(character(0))
}
\author{
Will Beasley
}
