% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash-and-salt.R
\name{hash_and_salt_sha_256}
\alias{hash_and_salt_sha_256}
\title{Salt and hash a value}
\usage{
hash_and_salt_sha_256(x, salt = "", min_characters = 1L,
  max_characters = 2048L)
}
\arguments{
\item{x}{A vector of values to convert.  it should be a \code{character} vector, or something that can be cast to a \code{character} vector.}

\item{salt}{A single-element character vector.}

\item{min_characters}{The minimum count of characters that \code{x} is allowed to be.  Must be an integer.}

\item{max_characters}{The maximum count of characters that \code{x} is allowed to be.  Must be an integer.}
}
\value{
A `character vector.
}
\description{
Uses \code{\link[digest:digest]{digest::digest()}} to hash as (salted) value, using 'SHA-256'.
If the \code{x} isn't already a \code{character} vector, it's converted to one (even if \code{x} inherits from character).
}
\examples{
x    <- letters[1:5]

salt <- "abc123"
hash_and_salt_sha_256(x, salt)

# If an unsalted hash is desired, leave it blank
hash_and_salt_sha_256(x)
}
\author{
Will Beasley
}
