% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-frame-metadata.R
\name{headstart_utilities}
\alias{column_class_headstart}
\alias{column_rename_headstart}
\alias{column_value_headstart}
\alias{headstart_utilities}
\title{Utilities for outputting characteristics of a datatset used it code.}
\usage{
column_rename_headstart( d )
column_class_headstart( d )
column_value_headstart( x )
}
\arguments{
\item{d}{A \code{data.frame} to describe.}

\item{x}{A vector to describe.}
}
\value{
Prints formatted code to the console.
}
\description{
These functions are used during the execution of a program.  Rather they produce snippets
that can be pasted into code, and help the developer avoid some typing.
}
\examples{
column_rename_headstart(datasets::OrchardSprays)
column_class_headstart(datasets::OrchardSprays)
column_value_headstart(datasets::OrchardSprays$treatment)
}
\author{
Will Beasley
}

