% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readr-spec-aligned.R
\name{readr_spec_aligned}
\alias{readr_spec_aligned}
\title{Align & qualify the de}
\usage{
readr_spec_aligned(...)
}
\arguments{
\item{...}{values passed to \code{\link[readr:spec_csv]{readr::spec_csv()}}.}
}
\value{
Prints formatted code to the console.
}
\description{
These functions are used during the execution of a program.  Rather they produce snippets
that can be pasted into code, and help the developer avoid some typing.
}
\examples{
readr_spec_aligned(system.file("test-data/subject-1.csv", package = "OuhscMunge"))
readr_spec_aligned(system.file("package-dependency-list.csv", package = "OuhscMunge"))
}
\author{
Will Beasley
}
