% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve-key-value.R
\name{retrieve_key_value}
\alias{retrieve_key_value}
\title{Read a value stored in a database.}
\usage{
retrieve_key_value(key, project_name, dsn_name, channel = NULL)
}
\arguments{
\item{key}{The key associated with the desired value.  Required character vecotr with one element}

\item{project_name}{The project name associated with the desired value.  Required character vecotr with one element}

\item{dsn_name}{Name of the locally-defined DSN passed to \url{RODBC::odbcConnect()}.}

\item{channel}{An \emph{optional} connection handle as returned by \code{\link[RODBC:odbcConnect]{RODBC::odbcConnect()}}.  See Details below. Optional.}
}
\value{
A \code{character} vector with one element.
}
\description{
Facilitates retrieval of key-value pairs that are stored in a SQL Server database.
}
\details{
The database table and stored procedure must defined as:\preformatted{CREATE TABLE [security_private].[tbl_key_value_static](
  [id] [smallint] IDENTITY(1,1) NOT NULL,
  [project] [varchar](50) NOT NULL,
  [attribute] [varchar](90) NOT NULL,
  [value] [varchar](200) NOT NULL,
  [file_last_updated_date] [date] NOT NULL,
  [retired] [bit] NOT NULL,
  CONSTRAINT [PK_tbl_key_value_static] PRIMARY KEY CLUSTERED
  (
    [id] ASC
  )WITH (PAD_INDEX = OFF, STATISTICS_NORECOMPUTE = OFF, IGNORE_DUP_KEY = OFF, ALLOW_ROW_LOCKS = ON, ALLOW_PAGE_LOCKS = ON) ON [PRIMARY]
) ON [PRIMARY]

CREATE PROCEDURE [security].[prc_key_value_static]
  @project varchar(50),
  @attribute varchar(90)
AS
BEGIN
	SET NOCOUNT ON;
	SELECT value from security_private.tbl_key_value_static
	WHERE project=@project AND attribute=@attribute
END
}
}
\note{
Currently only the 'static' key-value pairs are retrieved through this function.
Talk to Will if you need to retrieve the 'rolling' or the 'personal' key-value pairs.
}
\examples{
\dontrun{
value <- retrieve_key_value("file-server", "bbmc", "BbmcSecurity")
}
}
\author{
Will Beasley
}
