% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snake-case.R
\name{snake_case}
\alias{snake_case}
\title{Convert variable names to snake_case}
\usage{
snake_case(x)
}
\arguments{
\item{x}{A vector to of names to convert.}
}
\value{
A vector of converted names.
}
\description{
This function attempts to convert variables to snake_case, even if it's already in snake_case.  
The important regex lines were posted by Stack Overflow user \href{http://stackoverflow.com/users/129879/epost}{epost}
in \href{http://stackoverflow.com/a/1176023/1082435}{"Elegant Python function to convert CamelCase to snake_case?"}.
}
\examples{
snake_case(colnames(datasets::OrchardSprays))
snake_case(colnames(datasets::iris))
}
\author{
Will Beasley
}
