% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify-data-frame.R
\name{verify_data_frame}
\alias{verify_data_frame}
\title{Verify object inherits from \href{base::data.frame()}{data.frame}.}
\usage{
verify_data_frame(d, minimum_row_count = 10L)
}
\arguments{
\item{d}{The object to verify.  Required.}

\item{minimum_row_count}{The \code{data.frame} should have at least this may rows.  Defaults to 10.}
}
\description{
Check that the object inherits from \href{base::data.frame()}{data.frame}.
If not, throw an error.

A minimum row count can be used to check tha a trivially small number of rows was not returned.
If this is set to 0, the function is essentially \code{checkmate::assert_class(d, "data.frame")}.
}
\examples{
verify_data_frame(datasets::OrchardSprays, 20)
verify_data_frame(datasets::iris, 4)
}
\author{
Will Beasley
}
